Fix variable type in inttolongtype.java and add typeconvo.java for type conversion example

+++ Basicforjava/inttolongtype.java
@@ -4,7 +4,7 @@ import java.util.Scanner;
 public class inttolongtype {
     public static void main(String[] args) {
         int a = 100;
-        int b = a;
+        long b = a;
         System.out.println(b);
     }
 }
